flask_bot/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
blobfile==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
distlib==0.3.9
distro==1.9.0
Django==5.1.7
django-appconf==1.1.0
django-compressor==4.5.1
django-libsass==0.9
django-sass-processor==1.4.1
durationpy==0.9
et-xmlfile==1.1.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
libsass==0.23.0
llama_stack==0.2.4
llama_stack_client==0.2.4
load-dotenv==0.1.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.68.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.1
pillow==11.2.1
platformdirs==4.3.7
playwright==1.52.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyaml==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==13.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.2
rcssmin==1.1.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rjsmin==1.2.2
rpds-py==0.24.0
rsa==4.9
rubicon-objc==0.5.0
selenium==4.31.0
setuptools==80.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
sse-starlette==2.2.1
starlette==0.46.1
termcolor==3.0.1
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
